Added index.js entry point
